+++ src/loop_handler.py
@@ -12,7 +12,7 @@ from wrappers import last_fm_user
 from wrappers import system_tray_icon
 from wrappers import track_info
 
-logger = logging.getLogger("discord_fm").getChild(__name__)
+logger = logging.getLogger("fm_rpc").getChild(__name__)
 
 
 class LoopHandler: