+++ src/util/install/macos.py
@@ -5,10 +5,10 @@ from pathlib import Path
 from util import resource_path
 from util.install import BaseInstall
 
-logger = logging.getLogger("discord_fm").getChild(__name__)
+logger = logging.getLogger("fm_rpc").getChild(__name__)
 
 APP_PATH = Path("~/Applications/fm.rpc.app").expanduser()
-PLIST_NAME = "net.andro-32.discord_fm.launch.plist"
+PLIST_NAME = "net.andro-32.fm_rpc.launch.plist"
 LAUNCH_AGENTS_PATH = Path("~/Library/LaunchAgents").expanduser()
 PLIST_PATH = Path(LAUNCH_AGENTS_PATH, PLIST_NAME)