+++ src/wrappers/discord_rp.py
@@ -6,7 +6,7 @@ from pypresence import Presence, InvalidID, PipeClosed, ActivityType
 
 from wrappers import track_info
 
-logger = logging.getLogger("discord_fm").getChild(__name__)
+logger = logging.getLogger("fm_rpc").getChild(__name__)
 
 
 class DiscordRP: