+++ src/wrappers/last_fm_user.py
@@ -6,7 +6,7 @@ import pylast
 import util
 from wrappers import track_info
 
-logger = logging.getLogger("discord_fm").getChild(__name__)
+logger = logging.getLogger("fm_rpc").getChild(__name__)
 
 
 class LastFMUser: